refactor(private-credential): render-payload
